# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Resume Analyzer - Dual Mode Hiring Platform
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface CandidateAnalysisResult {
  /** Match score from 0 to 100 */
  matchScore: number;
  /** Keywords found in both resume and job description */
  matchedKeywords: string[];
  /** Keywords in job description but not in resume */
  missingKeywords: string[];
  /** AI-generated improvement suggestions */
  suggestions: string[];
  /** AI-identified resume strengths */
  strengths: string[];
  /** Text extracted from the PDF resume */
  extractedText: string;
}

/**
 * Overall hiring recommendation
 */
export type RecruiterEvaluationResultHiringRecommendation =
  (typeof RecruiterEvaluationResultHiringRecommendation)[keyof typeof RecruiterEvaluationResultHiringRecommendation];

export const RecruiterEvaluationResultHiringRecommendation = {
  Strong_Fit: "Strong Fit",
  Moderate_Fit: "Moderate Fit",
  Low_Fit: "Low Fit",
} as const;

/**
 * Recommended recruiter action
 */
export type RecruiterEvaluationResultNextStep =
  (typeof RecruiterEvaluationResultNextStep)[keyof typeof RecruiterEvaluationResultNextStep];

export const RecruiterEvaluationResultNextStep = {
  Shortlist_for_Screening: "Shortlist for Screening",
  Consider_with_Caution: "Consider with Caution",
  Not_Suitable_for_This_Role: "Not Suitable for This Role",
} as const;

export interface RecruiterEvaluationResult {
  /** Candidate fit score from 0 to 100 */
  fitScore: number;
  /** Overall hiring recommendation */
  hiringRecommendation: RecruiterEvaluationResultHiringRecommendation;
  /** Skills the candidate has that match the role */
  matchedSkills: string[];
  /** Required skills missing from the candidate profile */
  missingSkills: string[];
  /** Brief summary of candidate experience relevant to the role */
  experienceSummary: string;
  /** Candidate strengths relevant to the role */
  strengths: string[];
  /** Gaps or concerns about the candidate */
  concerns: string[];
  /** Recommended recruiter action */
  nextStep: RecruiterEvaluationResultNextStep;
  /** Brief explanation of the fit score */
  scoreRationale: string;
}

export interface ErrorResponse {
  error: string;
}

export type AnalyzeResumeBody = {
  /** PDF resume file */
  resume: Blob;
  /** Job description text */
  jobDescription: string;
};

export type EvaluateResumeBody = {
  /** Candidate PDF resume file */
  resume: Blob;
  /** Job description text */
  jobDescription: string;
};