# lib/api-zod/src/generated/types/candidateAnalysisResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Resume Analyzer - Dual Mode Hiring Platform
 * OpenAPI spec version: 0.1.0
 */

export interface CandidateAnalysisResult {
  /** Match score from 0 to 100 */
  matchScore: number;
  /** Keywords found in both resume and job description */
  matchedKeywords: string[];
  /** Keywords in job description but not in resume */
  missingKeywords: string[];
  /** AI-generated improvement suggestions */
  suggestions: string[];
  /** AI-identified resume strengths */
  strengths: string[];
  /** Text extracted from the PDF resume */
  extractedText: string;
}