# lib/api-zod/src/generated/types/recruiterEvaluationResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Resume Analyzer - Dual Mode Hiring Platform
 * OpenAPI spec version: 0.1.0
 */
import type { RecruiterEvaluationResultHiringRecommendation } from "./recruiterEvaluationResultHiringRecommendation";
import type { RecruiterEvaluationResultNextStep } from "./recruiterEvaluationResultNextStep";

export interface RecruiterEvaluationResult {
  /** Candidate fit score from 0 to 100 */
  fitScore: number;
  /** Overall hiring recommendation */
  hiringRecommendation: RecruiterEvaluationResultHiringRecommendation;
  /** Skills the candidate has that match the role */
  matchedSkills: string[];
  /** Required skills missing from the candidate profile */
  missingSkills: string[];
  /** Brief summary of candidate experience relevant to the role */
  experienceSummary: string;
  /** Candidate strengths relevant to the role */
  strengths: string[];
  /** Gaps or concerns about the candidate */
  concerns: string[];
  /** Recommended recruiter action */
  nextStep: RecruiterEvaluationResultNextStep;
  /** Brief explanation of the fit score */
  scoreRationale: string;
}